+++ crates/rustc_codegen_spirv/src/linker/simple_passes.rs
@@ -245,11 +245,11 @@ pub fn check_fragment_insts(sess: &Session, module: &Module) -> Result<()> {
                     .collect::<Vec<_>>()
                     .join("\n");
                 any_err = Some(
-                    sess.struct_err(&format!(
+                    sess.struct_err(format!(
                         "{} cannot be used outside a fragment shader",
                         inst.class.opname
                     ))
-                    .note(&note)
+                    .note(note)
                     .emit(),
                 );
             }